% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subchunkify.R
\name{subchunkify}
\alias{subchunkify}
\title{Erzeugt Subchunks für die Berichte}
\usage{
subchunkify(g, fig_height = 7, fig_width = 5, hide = FALSE)
}
\arguments{
\item{g}{Code (kann auch mit Aufzählung ("c(...)") benutzt werden)}

\item{fig_height}{A number.}

\item{fig_width}{A number.}

\item{hide}{description}
}
\value{
Einen Subchunk
}
\description{
Erzeugt Subchunks für die Berichte
}
\examples{
add(1, 1)
add(10, 1)
}
